fix: 删除无用文件，删除swc无用参数，插件开发时监听文件变化
