+++ packages/@plugin/plugin-babel/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "main": "index.ts",
   "scripts": {
-    "dev": "tsc",
+    "dev": "tsc -w",
     "dev:run": "tsc -w"
   },
   "dependencies": {
