+++ packages/@plugin/plugin-eslint/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "main": "index.ts",
   "scripts": {
-    "dev": "tsc"
+    "dev": "tsc -w"
   },
   "engines": {
     "node": ">=16.0.0"
