+++ packages/@plugin/plugin-pinia/package.json
@@ -4,7 +4,7 @@
   "version": "1.0.0",
   "description": "pinia plugin for vue",
   "scripts": {
-    "dev": "tsc"
+    "dev": "tsc -w"
   },
   "dependencies": {}
 }
