+++ packages/@plugin/plugin-swc/package.json
@@ -5,7 +5,7 @@
   "type": "module",
   "main": "index.ts",
   "scripts": {
-    "dev": "tsc"
+    "dev": "tsc -w"
   },
   "dependencies": {},
   "devDependencies": {}
