+++ packages/@plugin/plugin-typescript/package.json
@@ -5,7 +5,7 @@
   "type": "commonjs",
   "main": "index.ts",
   "scripts": {
-    "dev": "tsc"
+    "dev": "tsc -w"
   },
   "dependencies": {}
 }
