+++ packages/core/src/models/protocolGenerator/PluginToBuildToolAPI.ts
@@ -24,7 +24,8 @@ class PluginToBuildToolAPI extends ProtocolGeneratorAPI {
    * @param params.buildTool 构建工具选择
    */
   ADD_COMPILER_CONFIG(params) {
-    const { compiler, template, buildTool } = params;
+    const { compiler } = params;
+    const { template, buildTool } = this.props.preset;
     const buildToolConfigAst = this.props.buildToolConfigAst;
     // 默认生成的单独配置文件
     const compilerConfigMap = {
@@ -65,9 +66,11 @@ class PluginToBuildToolAPI extends ProtocolGeneratorAPI {
             },
           },
         ],
+        rules: [],
       }),
       rollup: ({ compiler }) => ({
         plugins: [require(`@rollup/plugin-${compiler}`)],
+        rules: [],
       }),
     };
